+++ source/fitz/font.c
@@ -388,11 +388,11 @@ void fz_drop_font_context(fz_context *ctx)
 	{
 		int i;
 
-		for (i = 0; i < nelem(ctx->font->base14); ++i)
+		for (i = 0; i < (int)nelem(ctx->font->base14); ++i)
 			fz_drop_font(ctx, ctx->font->base14[i]);
-		for (i = 0; i < nelem(ctx->font->cjk); ++i)
+		for (i = 0; i < (int)nelem(ctx->font->cjk); ++i)
 			fz_drop_font(ctx, ctx->font->cjk[i]);
-		for (i = 0; i < nelem(ctx->font->fallback); ++i)
+		for (i = 0; i < (int)nelem(ctx->font->fallback); ++i)
 		{
 			fz_drop_font(ctx, ctx->font->fallback[i].serif);
 			fz_drop_font(ctx, ctx->font->fallback[i].sans);
@@ -524,7 +524,7 @@ fz_font *fz_load_fallback_font(fz_context *ctx, int script, int language, int se
 	int subfont;
 	int size;
 
-	if (script < 0 || script >= nelem(ctx->font->fallback))
+	if (script < 0 || script >= (int)nelem(ctx->font->fallback))
 		return NULL;
 
 	/* TODO: bold and italic */
@@ -933,7 +933,7 @@ fz_new_cjk_font(fz_context *ctx, int ordering)
 {
 	const unsigned char *data;
 	int size, index;
-	if (ordering >= 0 && ordering < nelem(ctx->font->cjk))
+	if (ordering >= 0 && ordering < (int)nelem(ctx->font->cjk))
 	{
 		if (ctx->font->cjk[ordering])
 			return fz_keep_font(ctx, ctx->font->cjk[ordering]);
