+++ source/svg/svg-parse.c
@@ -140,7 +140,7 @@ svg_parse_transform(fz_context *ctx, svg_document *doc, const char *str, fz_matr
 		 */
 
 		keywordlen = 0;
-		while (svg_is_alpha(*str) && keywordlen < sizeof(keyword) - 1)
+		while (svg_is_alpha(*str) && keywordlen < (int)sizeof(keyword) - 1)
 			keyword[keywordlen++] = *str++;
 		keyword[keywordlen] = 0;
